[ci] Add check to ensure code generation remains in sync (#36)
* add --check for generated files and add clippy

* add CI

* add rustfmt

* fix up CI so it runs on only linux (windows fails)

* use git to check for uncommitted changes

* test

* Revert "test"

This reverts commit 2ca394776f89369d36d465b6fb5aa5e4700aaf9a.

+++ fearless_simd_gen/src/mk_fallback.rs
@@ -441,7 +441,7 @@ fn interleave_indices(
 ) -> TokenStream {
     let indices = {
         let indices = (0..len).collect::<Vec<_>>();
-        interleave(&indices, len / count as usize)
+        interleave(&indices, len / count)
     };
 
     make_list(indices.into_iter().map(func).collect::<Vec<_>>())
